Docstring cleanup + some renaming
