+++ src/ghga_connector/core/crypt/checksums.py
@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-"""TODO"""
+"""Wrapper functionality for checksum generation"""
 
 import hashlib
