+++ src/ghga_connector/core/crypt/decryption.py
@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-"""TODO"""
+"""This module contains Crypt4GH based decryption functionality"""
 
 from pathlib import Path
