+++ src/ghga_connector/core/crypt/encryption.py
@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-"""TODO"""
+"""This module contains Crypt4GH based encryption functionality"""
 
 import base64
 import os
