+++ src/ghga_connector/core/file_operations.py
@@ -21,7 +21,7 @@ from collections.abc import Iterator, Sequence
 from io import BufferedReader
 from pathlib import Path
 
-from ghga_connector.core.dataclasses import PartRange
+from ghga_connector.core.structs import PartRange
 
 
 def is_file_encrypted(file_path: Path):
