+++ src/ghga_connector/core/structs.py
