+++ src/ghga_connector/core/uploading/api_calls.py
@@ -27,7 +27,7 @@ from ghga_connector.core import exceptions
 from ghga_connector.core.constants import MAX_PART_NUMBER, TIMEOUT
 from ghga_connector.core.http_translation import ResponseExceptionTranslator
 from ghga_connector.core.uploading.abstract_uploader import UploaderBase
-from ghga_connector.core.uploading.dataclasses import UploadStatus
+from ghga_connector.core.uploading.structs import UploadStatus
 
 
 class Uploader(UploaderBase):
